Aeroport, passager ok

+++ src/main/java/org/odk/g1/pankuru/Service/Service/PaysService.java
@@ -1,5 +1,8 @@
 package org.odk.g1.pankuru.Service.Service;
 
+import org.odk.g1.pankuru.Entity.Humain.AdminCompagnie;
+import org.odk.g1.pankuru.Entity.Localite.Aeroport;
+import org.odk.g1.pankuru.Repository.HumainRepo.AdminCompagnieRepo;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -17,8 +20,14 @@ import lombok.AllArgsConstructor;
 public class PaysService implements CrudService<Pays, Integer>{
 
     private final PaysRepository paysRepository;
+    private final CompagnieService compagnieService;
+    private final UserService userService;
+    private final AdminCompagnieRepo adminCompagnieRepo;
     @Override
     public Pays ajout(Pays pays) {
+        Long adminCompagnieId = userService.getCurrentUsernameId();
+        Optional<AdminCompagnie> adminCompagnie = adminCompagnieRepo.findById(adminCompagnieId);
+        adminCompagnie.ifPresent(pays::setAdminCompagnie);
         return paysRepository.save(pays);
     }
 
@@ -29,25 +38,43 @@ public class PaysService implements CrudService<Pays, Integer>{
 
     @Override
     public Optional<Pays> trouverParId(Integer id) {
-        return paysRepository.findById(id);
+        Optional<Pays> pays = paysRepository.findById(id);
+        if (pays.isPresent()) {
+            if(compagnieService.getPaysByCompagnieId().contains(pays.get())){
+                return paysRepository.findById(id);
+            }
+        }
+        return Optional.empty();
     }
 
     @Override
     public Pays misAJour(Pays pays, Integer Id) {
-        Optional<Pays> paysExistant = paysRepository.findById(Id);
-        if (paysExistant.isPresent()) {
-            Pays paysAModifier = paysExistant.get();
-            paysAModifier.setNom(pays.getNom());
+        Optional<Pays> pays1 = paysRepository.findById(Id);
+        if (pays1.isPresent()) {
+            if(compagnieService.getPaysByCompagnieId().contains(pays1.get())){
+                Optional<Pays> paysExistant = paysRepository.findById(Id);
+                if (paysExistant.isPresent()) {
+                    Pays paysAModifier = paysExistant.get();
+                    paysAModifier.setNom(pays.getNom());
 
-            return paysRepository.save(paysAModifier);
-        } else {
-            throw new IllegalArgumentException("Le pays avec l'ID " + pays.getId() + "n'existe pas.");
+                    return paysRepository.save(paysAModifier);
+                } else {
+                    throw new IllegalArgumentException("Le pays avec l'ID " + pays.getId() + "n'existe pas.");
+                }
+            }
         }
+        return null;
     }
 
     @Override
     public void supprimer(Integer id) {
-        paysRepository.deleteById(id);
+        Optional<Pays> pays = paysRepository.findById(id);
+        if (pays.isPresent()) {
+            if(compagnieService.getPaysByCompagnieId().contains(pays.get())){
+                paysRepository.deleteById(id);
+            }
+        }
+
     }
     
 }
